refactor: write client file
