+++ packages/client/tsconfig.json
@@ -4,7 +4,7 @@
     "module": "CommonJS",
     "resolveJsonModule": true,
     "emitDecoratorMetadata": true,
-    "experimentalDecorators": true,
+    "experimentalDecorators": true
   },
   "include": [
     "src"
